[UPDATE] server 코드 개선
드라이브 찾는 로직 수정
로그 화면 가독성 개선

+++ Data/Data.h
@@ -33,6 +33,6 @@ public:
     int childLength;
     char child[500][MAX_PATH];
     char childAccessTime[500][100];
-    int childSize[500];
+    ULONGLONG childSize[500];
 };
